resume-parser/requirements.txt
fastapi
uvicorn
pdfminer.six
spacy
python-multipart
